pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "embex"
version = "0.1.0"
description = "Local-first CLI memory layer for AI agents — embed your codebase into a local vector database."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Embex"},
]

dependencies = [
    "typer>=0.9.0",
    "chromadb>=0.4.0",
    "openai>=1.0.0",
    "watchdog>=3.0.0",
    "tiktoken>=0.5.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "sentence-transformers>=2.2.0",
    "python-dotenv>=1.0.0",
    "mcp[cli]>=1.0.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tree-sitter-typescript>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
embex = "embex.cli:app"

[tool.setuptools.packages.find]
include = ["embex*"]
